requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
pandas>=2.2.0
numpy<2.0.0
torch==2.2.0+cpu
torchvision==0.17.0+cpu
mlflow>=2.10.0
prometheus-fastapi-instrumentator>=6.1.0
pytest>=7.4.2
pypdf>=3.15.5
python-multipart>=0.0.6
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
Pillow>=10.1.0
